climbing: refactor - move to services/tagging/climbing
