+++ src/components/FeaturePanel/Climbing/ClimbingGradesTable/ClimbingGradesTableSettings.tsx
@@ -1,7 +1,7 @@
 import { useUserSettingsContext } from '../../../utils/UserSettingsContext';
 import { Alert, Box, Chip, Stack, Typography } from '@mui/material';
 import { t } from '../../../../services/intl';
-import { GRADE_SYSTEMS } from '../../../../services/tagging/climbing';
+import { GRADE_SYSTEMS } from '../../../../services/tagging/climbing/gradeSystems';
 import React from 'react';
 import { useGradeSystemsStatus } from '../utils/useVisibleGradeSystems';
